+++ lib/fastlane/plugin/wpmreleasetoolkit/actions/android/android_build_prechecks.rb
@@ -10,7 +10,6 @@ module Fastlane
         # Verify that the current branch is a release branch. Notice that `ensure_git_branch` expects a RegEx parameter
         ensure_git_branch(branch: '^release/') unless other_action.is_ci
 
-        has_alpha_version = params[:has_alpha_version]
         project_root_folder = params[:project_root_folder]
         project_name = params[:project_name]
         build_gradle_path = params[:build_gradle_path] || (File.join(project_root_folder || '.', project_name, 'build.gradle') unless project_name.nil?)
@@ -20,15 +19,13 @@ module Fastlane
         unless !params[:beta] && !params[:final]
           beta_version = Fastlane::Helper::Android::VersionHelper.get_release_version(
             build_gradle_path: build_gradle_path,
-            version_properties_path: version_properties_path,
-            has_alpha_version: has_alpha_version
+            version_properties_path: version_properties_path
           )
         end
         if params[:alpha]
           alpha_version = Fastlane::Helper::Android::VersionHelper.get_alpha_version(
             build_gradle_path: build_gradle_path,
-            version_properties_path: version_properties_path,
-            has_alpha_version: has_alpha_version
+            version_properties_path: version_properties_path
           )
         end
 
@@ -98,7 +95,6 @@ module Fastlane
                                                                project_root_folder]),
           Fastlane::Helper::Deprecated.project_root_folder_config_item,
           Fastlane::Helper::Deprecated.project_name_config_item,
-          Fastlane::Helper::Deprecated.has_alpha_version_config_item,
         ]
       end