+++ lib/fastlane/plugin/wpmreleasetoolkit/actions/android/android_bump_version_beta.rb
@@ -10,7 +10,6 @@ module Fastlane
         # Verify that the current branch is a release branch. Notice that `ensure_git_branch` expects a RegEx parameter
         ensure_git_branch(branch: '^release/')
 
-        has_alpha_version = params[:has_alpha_version]
         project_root_folder = params[:project_root_folder]
         project_name = params[:project_name]
         build_gradle_path = params[:build_gradle_path] || (File.join(project_root_folder || '.', project_name, 'build.gradle') unless project_name.nil?)
@@ -18,13 +17,11 @@ module Fastlane
 
         current_version = Fastlane::Helper::Android::VersionHelper.get_release_version(
           build_gradle_path: build_gradle_path,
-          version_properties_path: version_properties_path,
-          has_alpha_version: has_alpha_version
+          version_properties_path: version_properties_path
         )
         current_version_alpha = Fastlane::Helper::Android::VersionHelper.get_alpha_version(
           build_gradle_path: build_gradle_path,
-          version_properties_path: version_properties_path,
-          has_alpha_version: has_alpha_version
+          version_properties_path: version_properties_path
         )
         new_version_beta = Fastlane::Helper::Android::VersionHelper.calc_next_beta_version(current_version, current_version_alpha)
         new_version_alpha = current_version_alpha.nil? ? nil : Fastlane::Helper::Android::VersionHelper.calc_next_alpha_version(new_version_beta, current_version_alpha)
@@ -40,8 +37,7 @@ module Fastlane
         Fastlane::Helper::Android::VersionHelper.update_versions(
           new_version_beta,
           new_version_alpha,
-          version_properties_path: version_properties_path,
-          has_alpha_version: has_alpha_version
+          version_properties_path: version_properties_path
         )
         UI.message 'Done!'
 
@@ -81,7 +77,6 @@ module Fastlane
                                                                project_root_folder]),
           Fastlane::Helper::Deprecated.project_root_folder_config_item,
           Fastlane::Helper::Deprecated.project_name_config_item,
-          Fastlane::Helper::Deprecated.has_alpha_version_config_item,
         ]
       end