+++ lib/fastlane/plugin/wpmreleasetoolkit/actions/android/android_bump_version_hotfix.rb
@@ -6,7 +6,6 @@ module Fastlane
 
         require_relative '../../helper/android/android_git_helper'
 
-        has_alpha_version = params[:has_alpha_version]
         project_root_folder = params[:project_root_folder]
         project_name = params[:project_name]
         build_gradle_path = params[:build_gradle_path] || (File.join(project_root_folder || '.', project_name, 'build.gradle') unless project_name.nil?)
@@ -16,8 +15,7 @@ module Fastlane
 
         current_version = Fastlane::Helper::Android::VersionHelper.get_release_version(
           build_gradle_path: build_gradle_path,
-          version_properties_path: version_properties_path,
-          has_alpha_version: has_alpha_version
+          version_properties_path: version_properties_path
         )
         new_version = Fastlane::Helper::Android::VersionHelper.calc_next_hotfix_version(params[:version_name], params[:version_code]) # NOTE: this just puts the name/code values in a tuple, unchanged (no actual calc/bumping)
         new_release_branch = "release/#{params[:version_name]}"
@@ -32,8 +30,7 @@ module Fastlane
         Fastlane::Helper::Android::VersionHelper.update_versions(
           new_version,
           nil,
-          version_properties_path: version_properties_path,
-          has_alpha_version: has_alpha_version
+          version_properties_path: version_properties_path
         )
         UI.message 'Done!'
 
@@ -87,7 +84,6 @@ module Fastlane
                                                                project_root_folder]),
           Fastlane::Helper::Deprecated.project_root_folder_config_item,
           Fastlane::Helper::Deprecated.project_name_config_item,
-          Fastlane::Helper::Deprecated.has_alpha_version_config_item,
         ]
       end