+++ lib/fastlane/plugin/wpmreleasetoolkit/helper/android/android_version_helper.rb
@@ -30,11 +30,10 @@ module Fastlane
         #         - If this version is a hotfix (more than 2 parts and 3rd part is non-zero), returns the "X.Y.Z" formatted string
         #         - Otherwise (not a hotfix / 3rd part of version is 0), returns "X.Y" formatted version number
         #
-        def self.get_public_version(build_gradle_path:, version_properties_path:, has_alpha_version:)
+        def self.get_public_version(build_gradle_path:, version_properties_path:)
           version = get_release_version(
             build_gradle_path: build_gradle_path,
-            version_properties_path: version_properties_path,
-            has_alpha_version: has_alpha_version
+            version_properties_path: version_properties_path
           )
           vp = get_version_parts(version[VERSION_NAME])
           return "#{vp[MAJOR_NUMBER]}.#{vp[MINOR_NUMBER]}" unless is_hotfix?(version)
@@ -46,10 +45,10 @@ module Fastlane
         #
         # @return [Hash] A hash with 2 keys "name" and "code" containing the extracted version name and code, respectively
         #
-        def self.get_release_version(build_gradle_path:, version_properties_path:, has_alpha_version:)
+        def self.get_release_version(build_gradle_path:, version_properties_path:)
           return get_version_from_properties(version_properties_path: version_properties_path) if File.exist?(version_properties_path)
 
-          section = has_alpha_version.nil? ? 'defaultConfig' : 'vanilla {'
+          section = 'defaultConfig'
           name = get_version_name_from_gradle_file(build_gradle_path, section)
           code = get_version_build_from_gradle_file(build_gradle_path, section)
           { VERSION_NAME => name, VERSION_CODE => code }
@@ -76,13 +75,10 @@ module Fastlane
 
         # Extract the version name and code from the `version.properties` file in the project root
         #
-        # @return [Hash] A hash with 2 keys `"name"` and `"code"` containing the extracted version name and code, respectively,
-        #                or `nil` if `$HAS_ALPHA_VERSION` is not defined.
+        # @return [Hash] A hash with 2 keys `"name"` and `"code"` containing the extracted version name and code, respectively
         #
-        def self.get_alpha_version(build_gradle_path:, version_properties_path:, has_alpha_version:)
+        def self.get_alpha_version(build_gradle_path:, version_properties_path:)
           return get_version_from_properties(version_properties_path: version_properties_path, is_alpha: true) if File.exist?(version_properties_path)
-
-          return nil if has_alpha_version.nil?
 
           section = 'defaultConfig'
           name = get_version_name_from_gradle_file(build_gradle_path, section)
@@ -278,12 +274,11 @@ module Fastlane
         #
         # @return [String] The next release version name to use after bumping the currently used release version.
         #
-        def self.bump_version_release(build_gradle_path:, version_properties_path:, has_alpha_version:)
+        def self.bump_version_release(build_gradle_path:, version_properties_path:)
           # Bump release
           current_version = get_release_version(
             build_gradle_path: build_gradle_path,
-            version_properties_path: version_properties_path,
-            has_alpha_version: has_alpha_version
+            version_properties_path: version_properties_path
           )
           UI.message("Current version: #{current_version[VERSION_NAME]}")
           new_version = calc_next_release_base_version(current_version)
@@ -296,7 +291,7 @@ module Fastlane
         # @param [Hash] new_version_beta The version hash for the beta, containing values for keys "name" and "code"
         # @param [Hash] new_version_alpha The version hash for the alpha , containing values for keys "name" and "code"
         #
-        def self.update_versions(new_version_beta, new_version_alpha, version_properties_path:, has_alpha_version:)
+        def self.update_versions(new_version_beta, new_version_alpha, version_properties_path:)
           if File.exist?(version_properties_path)
             replacements = {
               versionName: (new_version_beta || {})[VERSION_NAME],
@@ -312,7 +307,7 @@ module Fastlane
             end
             File.write(version_properties_path, content)
           else
-            update_version(new_version_beta, has_alpha_version.nil? ? 'defaultConfig' : 'vanilla {')
+            update_version(new_version_beta, 'defaultConfig')
             update_version(new_version_alpha, 'defaultConfig') unless new_version_alpha.nil?
           end
         end